fix(demo): improve worker config
