+++ demo/src/tools/fakeWorker.ts
@@ -1,0 +1,4 @@
+export class Worker {
+  constructor (public url: string | URL, public options?: WorkerOptions) {
+  }
+}
